Update header and about
Home page header
Last line, full stop after ‘Build capability’ then capital D for ‘deliver’.
Under this, “have questions now…..” piece - everything has a capital letter.  It shouldn’t

About

Title here currently ‘Enabling your path to change…’ subtitle ‘Making the most of your business’ can we replace these?
In place of these have - top title ‘Arabica Transformation Consulting’ and under it, ‘Agility from the boardroom to the customer’

Add text under this:
At Arabica Transformation we are passionate about helping organisations evolve and transform to success in today’s challenging business environments. We are a group of highly experienced transformation leaders and practitioners who specialise in supporting organisations that are looking to transform their business to take advantage of lean and agile ways of working. 

Our experience has shown us that a piecemeal approach to transformation support is ineffective, as is a multi-supplier approach that separates leadership, strategy and delivery. We have also found that using many independent individuals results in poor coordination and lack of consistency. 

At Arabica Transformation we are a group of like-minded transformation professionals who share the same philosophy, approach and practices. We are capable of supporting the entire scope of your organisational transformation from the boardroom to the customer.